+++ packages/eventbridge/CHANGELOG.md
@@ -1,7 +1,7 @@
-# @tayori/eventbridge
+# @kotodayori/eventbridge
 
 ## 0.0.1
 
 ### Minor Changes
 
-- Initial release of `@tayori/eventbridge`: AWS EventBridge adapter for Tayori.
+- Initial release of `@kotodayori/eventbridge`: AWS EventBridge adapter for Kotodayori.